Separate ChatApplication class from main file
